+++ src/pages/user/switch/index.js
@@ -1,29 +1,35 @@
 import Taro,{Component} from '@tarojs/taro';
 import {View} from '@tarojs/components';
 import './index.scss';
-import jump from './../../utils/jump';
-import Event from './../../utils/event';
-let myEvent = new Event();
+import store from './../../../store/';
+import {getChangeUserAction} from './../../../store/actionCreators';
 
 export default class ChangeUser extends Component{
-
-    constructor(){
+    
+    constructor(props){
         super(...arguments);
-
-        this.state = {
-            isAgent:false
-        }
+        this.state = store.getState();
+        this.handleChangeState = this.handleChangeState.bind(this);
+        this.handleStoreChange = this.handleStoreChange.bind(this);
+        store.subscribe(this.handleStoreChange)
+        console.log(store.getState());
     }
 
-    switchUser(){
-        console.log("fdsafd");
-        myEvent.emit("changeUser");
+    handleChangeState(){
+        const action = getChangeUserAction(!this.state.isAgent);
+        store.dispatch(action);
+        console.log("handleChangeState");
+    }
+
+    handleStoreChange(){
+        this.setState(store.getState());
+        console.log("store change");
     }
 
     render(){
         return (
-          <View className="mp-user-changeuser" onClick={this.switchUser.bind(this)}> 
-            切换为用户
+          <View className="mp-user-changeuser" onClick={this.handleChangeState.bind(this)}> 
+                切换为用户 {this.state.isAgent}
           </View>
         )
     }