+++ src/store/actionCreators.js
@@ -1,0 +1,6 @@
+import {CHANGE_USER} from './actionTypes';
+
+export const getChangeUserAction = (value) => ({
+    type: CHANGE_USER,
+    value
+})