+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * 新增活动页面.
 * 新增广告预览页面.
 * 新增订单列表.   by shawn 2019年04月06日
+* 引入 redux.
 
 ### Bug Fixes
 * 
@@ -19,4 +20,21 @@
 
 ### new 
 * css flex 知识 属性 : https://www.cnblogs.com/ghfjj/p/6529733.html
-* 
+* taro 映入 redux
+* reducer 能达到之前的数据和当前的信息做一些处理,返回新的state.
+
+```
+1、如何创建 store.
+2、如何创建 reducer.
+3、redux 流程图.
+4、action 和 reducer.
+5、引入 actionTypes.
+6、actionCreator 统一创建 action. 提高代码可维护性.
+7、store 是唯一的.
+8、store 只能自己改变自己的内容.
+9、reducer 是纯函数.
+redux 核心api.
+store.dispatch.
+store.getState
+store.subscribe
+```